requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
botocore==1.23.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.7
click==8.0.3
cloudinary==1.30.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-cleanup==6.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.9.0
django-filter==21.1
django-on-heroku==1.1.2
django-templated-mail==1.1.1
djangochannelsrestframework==1.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
dynaconf==3.1.7
gunicorn==20.1.0
hvac==0.11.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
loguru==0.5.3
MarkupSafe==2.0.1
numpy==1.22.1
oauthlib==3.1.1
pandas==1.4.0
Pillow==8.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
service-identity==21.1.0
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
whitenoise==6.2.0
zope.interface==5.4.0
